Multiple quality of life improvements (#89)
- Remove usage of reflection
- add .env to .gitignore
- add support for extended resonance filter range in IT
- clean up and fill out effect player interfaces
- upgrade voice package to take advantage of more generics support
- remove stupid usages of generics
- split out structures from player interfaces

+++ internal/voice/pcm.go
@@ -319,7 +319,7 @@ func (v *pcmVoice) IsFilterEnvelopeEnabled() bool {
 	return v.filterEnv.IsEnabled()
 }
 
-func (v *pcmVoice) GetCurrentFilterEnvelope() float32 {
+func (v *pcmVoice) GetCurrentFilterEnvelope() int8 {
 	return v.filterEnv.GetCurrentValue()
 }
